add: fill network options
